+++ src/main/webapp/resources/js/apis/projects/samples.js
@@ -1,36 +1,58 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import { setBaseUrl } from "../../utilities/url-utilities";
+import { getProjectIdFromUrl, setBaseUrl } from "../../utilities/url-utilities";
 
-const URL = setBaseUrl(`/ajax/projects/${window.project.id}/samples`);
+const PROJECT_ID = getProjectIdFromUrl();
+const URL = setBaseUrl(`/ajax/projects/${PROJECT_ID}/samples`);
 
 /**
  * Redux API for handling project samples queries.
  * @type {Api<(args: (string | FetchArgs), api: BaseQueryApi, extraOptions: {}) => MaybePromise<QueryReturnValue<unknown, {status: number, data: unknown} | {status: "FETCH_ERROR", data?: undefined, error: string} | {status: "PARSING_ERROR", originalStatus: number, data: string, error: string} | {status: "CUSTOM_ERROR", data?: unknown, error: string}, FetchBaseQueryMeta>>, {getSampleIdsForProject: *}, string, never, typeof coreModuleName> | Api<(args: (string | FetchArgs), api: BaseQueryApi, extraOptions: {}) => MaybePromise<QueryReturnValue<unknown, {status: number, data: unknown} | {status: "FETCH_ERROR", data?: undefined, error: string} | {status: "PARSING_ERROR", originalStatus: number, data: string, error: string} | {status: "CUSTOM_ERROR", data?: unknown, error: string}, FetchBaseQueryMeta>>, {getSampleIdsForProject: *}, string, never, any>}
  */
 export const samplesApi = createApi({
-  reducerPath: `samplesApi`,
+  reducerPath: "samplesApi",
   baseQuery: fetchBaseQuery({
-    baseUrl: setBaseUrl(`/ajax/samples`),
+    baseUrl: URL
   }),
-  endpoints: (build) => ({
-    getSampleIdsForProject: build.query({
-      query: (projectId) => ({
-        url: `identifiers?projectId=${projectId}`,
-      }),
+  endpoints: builder => ({
+    listSamples: builder.query({
+      query: body => ({ method: "POST", body })
     }),
-    shareSamplesWithProject: build.mutation({
-      query: (body) => ({
-        url: `share`,
+    merge: builder.mutation({
+      query: ({ projectId, request }) => ({
+        url: "/merge",
+        method: "POST",
+        body: request
+      })
+    }),
+    remove: builder.mutation({
+      query: sampleIds => ({
+        url: "/remove",
+        method: "DELETE",
+        body: { sampleIds }
+      })
+    }),
+    shareSamplesWithProject: builder.mutation({
+      query: body => ({
+        url: `/share`,
         method: `POST`,
-        body,
-      }),
+        body
+      })
     }),
-  }),
+    //TODO: This should not be in the slice but async thunk (update in metadata security)
+    getSampleIdsForProject: builder.query({
+      query: projectId => ({
+        url: `/identifiers?id=${projectId}`
+      })
+    })
+  })
 });
 
 export const {
+  useListSamplesQuery,
+  useMergeMutation,
+  useRemoveMutation,
   useGetSampleIdsForProjectQuery,
-  useShareSamplesWithProjectMutation,
+  useShareSamplesWithProjectMutation
 } = samplesApi;
 
 /**
@@ -55,9 +77,9 @@ export async function createNewSample({ name, organism }) {
   const response = await fetch(setBaseUrl(`${URL}/add-sample`), {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
+      "Content-Type": "application/json"
     },
-    body: JSON.stringify({ name: name.trim(), organism }),
+    body: JSON.stringify({ name: name.trim(), organism })
   });
 
   return response;
@@ -67,13 +89,28 @@ export async function shareSamplesWithProject({
   sampleIds,
   targetId,
   locked,
-  remove,
+  remove
 }) {
   return await fetch(setBaseUrl(`ajax/samples/share`), {
     method: "POST",
     headers: {
-      "Content-Type": "application/json",
+      "Content-Type": "application/json"
     },
-    body: JSON.stringify({ currentId, sampleIds, targetId, locked, remove }),
+    body: JSON.stringify({ currentId, sampleIds, targetId, locked, remove })
   });
+}
+
+/**
+ * Get get minimal information for all samples in a project.
+ * @param {object} options - current table filters
+ * @returns {Promise<*>}
+ */
+export async function getMinimalSampleDetailsForFilteredProject(options) {
+  return await fetch(`${URL}/ids`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json"
+    },
+    body: JSON.stringify(options)
+  }).then(response => response.json());
 }